requirements.txt
# local package
-e .

# external requirements
click>=8.0.1
Sphinx>=4.1.2
coverage>=5.5
flake8>=3.9.2
python-dotenv>=0.5.1
myst-parser>=0.15.1
pyyaml


# Project requirements
pandas
numpy
seaborn
sklearn
matplotlib
requests
